requirements.txt
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
PyJWT==2.3.0
python-dotenv==0.19.2
